Backend/requirements.txt
asgiref==3.10.0
Django==5.1.2
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
pillow==10.4.0
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
uritemplate==4.2.0
